make remove after S3 upload optional

+++ open_lm/datapreprocess/make_2048.py
@@ -48,11 +48,14 @@ def write_to_shard(chunks, shard_writer):
         shard_writer.write({"__key__": f"{idx:012d}", "txt": str(chunk)})
 
 
-def upload_to_s3_and_remove(fname, s3_base=S3_BASE):
+def upload_to_s3_and_remove(fname, s3_base=S3_BASE, remove=True):
     """Uploads file to s3 and removes it from local file system"""
     fname_split = fname.split("/")
     s3_path = s3_base.replace('/', '') + "/"  + fname_split[-2].replace('/', '') + "/" + fname_split[-1].replace('/', '')
-    cmd = f"aws s3 cp {fname} {s3_path} && rm {fname}"
+    if remove:
+        cmd = f"aws s3 cp {fname} {s3_path} && rm {fname}"
+    else:
+        cmd = f"aws s3 cp {fname} {s3_path}"   
     print("COMMAND:", cmd)
     os.system(cmd)
 
@@ -131,7 +134,7 @@ def process_files(file_list, buffer, enc, buffer_lock):
                                 buffer.append(chunk)
 
 
-def consumer(my_id, output_dir, threads, buffer, buffer_lock, num_consumers, upload_to_s3=False, s3_base_dir=S3_BASE):
+def consumer(my_id, output_dir, threads, buffer, buffer_lock, num_consumers, upload_to_s3=False, s3_base_dir=S3_BASE, remove=True):
     output_directory = f"{output_dir}/{CHUNK_SIZE - 1}-v1/{my_id}"
     os.makedirs(output_directory, exist_ok=True)
     shard_writer = ShardWriter(os.path.join(output_directory, "shard-%07d.tar"), maxcount=SHARD_SIZE)
@@ -153,7 +156,7 @@ def consumer(my_id, output_dir, threads, buffer, buffer_lock, num_consumers, upl
             print(f"I am {my_id} and I am writing a shard.", len(buffer))
             write_to_shard(chunks, shard_writer)
             if upload_to_s3:
-                upload_to_s3_and_remove(shard_writer.fname, s3_base=s3_base_dir)
+                upload_to_s3_and_remove(shard_writer.fname, s3_base=s3_base_dir, remove=remove)
             # print("FNAME", shard_writer.fname)
             chunks = []
             time_for_shard = time.time() - start_time
@@ -174,7 +177,7 @@ def consumer(my_id, output_dir, threads, buffer, buffer_lock, num_consumers, upl
 
         write_to_shard(chunks, shard_writer)
         if upload_to_s3:
-            upload_to_s3_and_remove(shard_writer.fname, s3_base=s3_base_dir)
+            upload_to_s3_and_remove(shard_writer.fname, s3_base=s3_base_dir, remove=remove)
         chunks = []
 
 
@@ -193,7 +196,8 @@ def main(
     num_workers=32,
     num_consumers=8,
     upload_to_s3=False,
-    s3_base_dir=S3_BASE
+    s3_base_dir=S3_BASE,
+    remove=False
 ):
     os.makedirs(f"{output_dir}/tars-{CHUNK_SIZE - 1}-v1", exist_ok=True)
     input_files = [glob.glob(input_file) for input_file in input_files]
@@ -240,6 +244,7 @@ def main(
                 num_consumers,
                 upload_to_s3,
                 s3_base_dir,
+                remove
             ),
         )
         t.start()
@@ -254,6 +259,7 @@ if __name__ == "__main__":
     parser.add_argument("--num-consumers", type=int, default=8)
     parser.add_argument("--upload-to-s3", action="store_true")
     parser.add_argument("--s3-base-dir", type=str, default="s3://llched-raw/open_lm_run_processed_data/export/")
+    parser.add_argument("--remove", action="store_true")
 
     args = parser.parse_args()
 
@@ -264,4 +270,5 @@ if __name__ == "__main__":
         args.num_consumers,
         args.upload_to_s3,
         args.s3_base_dir,
+        args.remove
     )
